Flatten domain_ops into per-entity manager modules; remove resolve.rs
The domain_ops/ directory added no concept of its own — its per-entity
files were already the manager's entity surface. Lift each up to
manager/<entity>.rs and absorb the remaining I/O resolvers into their
entity homes: find_release_detail_with + cover_ref_for + release_file_cloud_key
onto release.rs, resolve_album_detail onto album.rs,
playback_info_from_track_release onto track.rs, the storage-page DB ops
into the existing storage.rs. The cross-entity delete helpers
(queue_release_files_for_deletion, queue_release_cover_for_deletion,
cover_cloud_key) stay private but move to manager/mod.rs, the parent both
release and album reach — the same reachability domain_ops/mod.rs gave them.
