+++ bae-core/src/library/manager/artist.rs
