+++ bae-core/src/library/manager/image.rs
